--- src/fileUtils.test.js
@@ -1,5 +1,0 @@
-describe('fileUtils test cases', () => {
-    it('To test jest setup', () => {
-        expect(true).toBeTruthy()
-    })
-})
